Added logging info for TradeAmendmentTemplate. Previously there was zero logging. Trivial renamings. Corrected return type of ExtendedTrade#nextTradeLegId

+++ trade-publisher/src/main/java/io/alw/css/tradepublisher/template/domain/FxTrade.java
@@ -18,7 +18,7 @@ public final class FxTrade implements ExtendedTrade {
     }
 
     @Override
-    public int nextTradeLegId() {
+    public long nextTradeLegId() {
         return ++nextTradeLegId;
     }
